Made web stuff run in a thread.
Lucas, please read:
- For the portals: Will there be 2 per player or 2 total? If total, we
need to restucture where they are stored.

- I had to subclass EntityPlayer into EntityPlayerWebControlled so that
the web controlled player didn't listen to the keyboard. This means it
isn't currently following through on its jums, etc. We need to clean up
EntityPlayer so that the subclass can override the human player input
only, and not all of the other stuff.

- There is a weird effect when running into walls that the animation
goes over the wall. Can we fix this?

- This game, with a small screen, is using ~1gb of ram on my computer.
Can we fix that?

- Do not run multiplayer yet. The way it is set up, it burns through my
alloted free file operations, and we have ~20 minutes/day playing the
game multiplayer. Working on solutions, but for now we are stuck with
that.