Got Portals to work!
They now are angled the right way, they are actually on the wall and
not floating, and stuff!
AND: made it so when you successfully shoot a blue / orange portal, it
will switch which portal you shoot! Yay!
AND: Added a colored box int the bottom left-hand corner that indicates
the color that will be shot.