Added the MessageHookInterface
The MessageHookInterface can be used to have an object receive updates
when a message is received.
The hook method MessageHook(Message m) will receive a Message object,
from which the b64 encoded data can be received along with the type of
message it is.

If the object implementing the MessageHookInterface is the proper one
for that message, it should return True so that the message isn't
repeated to the other hooks.

The hook should be registered when the object is constructed.

In general, the same objects that accept the hook should also be able to
create one. This is done by creating a Message object and passing it to
MessageHolder.addMessageToSend(Message m). From here, the message will
be passed to the other player, who should have registered a hook to
receive it.